rename valideringsfil
